requirements.txt
fastapi
uvicorn 
websockets 
python-dotenv 
twilio
openai
google-api-python-client
google-auth
google-auth-httplib2
python-dateutil